+++ CHANGES.md
@@ -7,6 +7,7 @@ CHANGES
 * Add fullscreen enter / leave events.
 * Add compositionstart event.
 * Add clibpoard paste event.
+* Tab close event now have an option to ignore itself when closing window.
 
 ## 3.2.0 (2025-06-20)